+++ src/main/java/spaska/classifiers/OneR.java
@@ -85,7 +85,7 @@ public class OneR implements IClassifier {
 		// set portion of instances reaching a condition
 		int listCounter = 0;
 		for (Condition c : conditions) {
-			c.setReach(result.get(listCounter).size() / totalKnown);
+			c.setReach(((double)result.get(listCounter).size()) / totalKnown);
 			listCounter++;
 		}
 		// handle unknown
